testing: attempt to fix spot detection tests
