+++ software/tests/data/laser_af_camera.png
